feat: Add Go module and dependency for mux
This commit adds a Go module and a dependency for the mux package. The module is created for the "busca-cep-mux" package, and the dependency is added for the "github.com/gorilla/mux" package with version 1.8.1. This will enable the use of the mux package for handling HTTP requests and routing in the "busca-cep-mux" application.