feat: Add HTTP client functionality
This commit adds functionality to make HTTP requests using the `net/http` package in Go. It includes two new files, `main.go` and `main_test.go`, that demonstrate how to use the `http.Client` struct to send GET and POST requests. The GET request retrieves the body of a webpage from "http://google.com" and prints it to the console. The POST request sends a JSON payload to "http://google.com" and copies the response body to the standard output. This will allow users to easily perform HTTP requests in their Go programs.