feat: Refactor HTTP client code and add examples
This commit refactors the HTTP client functionality to improve code readability and maintainability. It also adds examples to demonstrate how to use the HTTP client to send GET and POST requests in different scenarios. This will make it easier for users to understand and utilize the HTTP client feature in their Go programs.